Move Popover, Tag, and Tooltip components to DataDisplay directory
